+++ packages/react-server/src/features/error/shared.ts
