+++ packages/react-server/src/features/router/server.tsx
@@ -1,6 +1,6 @@
 import { sortBy } from "@hiogawa/utils";
 import React from "react";
-import { type ReactServerErrorContext, createError } from "../../lib/error";
+import { type ReactServerErrorContext, createError } from "../error/shared";
 import { renderMetadata } from "../meta/server";
 import type { Metadata } from "../meta/utils";
 import type { ApiRouteMoudle } from "./api-route";
