+++ packages/react-server/src/runtime-client.ts
@@ -4,7 +4,7 @@ export {
   ErrorBoundary,
   RedirectBoundary,
   NotFoundBoundary,
-} from "./lib/client/error-boundary";
+} from "./features/error/error-boundary";
 export {
   LayoutContent,
   LayoutMatchProvider,
